Server-side template attempt
There is not enough information client-side to properly
parse out "dataset" or "screen" from the path. Bio-Formats
needs to have already run on the files and declared the
number of omitted levels.

As a solution, this attempts to push the parsing of the
file paths server-side. (Additionally, it moves to a regex-
based representation).

This however does not work as expected since the import
mechanism strips off all directories. Likely we don't
have much choice but to use the clientPath field.